feat: split risk-factors
