+++ components/risk-factors.js
@@ -1,67 +1,12 @@
 import React, {useState} from 'react'
 import PropTypes from 'prop-types'
 
-function RadioChoices({icon, choices, name, value, onChange, children}) {
-  return (
-    <>
-      <div
-        style={{
-          display: 'flex',
-          flexDirection: 'row',
-          paddingBottom: 0,
-          marginBottom: 0
-        }}
-      >
-        <div style={{flex: '0 1 40px'}}>
-          <i className={`fas ${icon}`} aria-hidden='true' />
-        </div>
-        <div
-          style={{
-            flex: '1 1 auto',
-            paddingBottom: 0,
-            marginBottom: 0
-          }}
-        >
-          <label>{children}</label>
-        </div>
-      </div>
-      <div
-        style={{
-          display: 'flex',
-          flexDirection: 'row'
-        }}
-      >
-        <div style={{flex: '0 1 60px'}} />
-        {choices.map(choice => (
-          <label key={choice.title} style={{margin: 0, padding: 0}}>
-            <input
-              name={name}
-              type='radio'
-              value={choice.value}
-              checked={value === choice.value}
-              onChange={() => onChange(choice.value)}
-            />{' '}
-            {choice.title}
-          </label>
-        ))}
-      </div>
-    </>
-  )
-}
-
 function RiskFactors({handleRiskFactors}) {
   const [breathingDisease, setBreathingDisease] = useState(false)
   const [cancer, setCancer] = useState(false)
   const [diabetes, setDiabetes] = useState(false)
-  const [heartDisease, setHeartDisease] = useState(false)
-  const [immunosuppressantDisease, setImmunosuppressantDisease] = useState(false)
-  const [immunosuppressantDrug, setImmunosuppressantDrug] = useState(false)
-  const [heartDiseaseAlgo, setHeartDiseaseAlgo] = useState(false)
-  const [immunosuppressantDiseaseAlgo, setImmunosuppressantDiseaseAlgo] = useState(false)
-  const [immunosuppressantDrugAlgo, setImmunosuppressantDrugAlgo] = useState(false)
   const [kidneyDisease, setKidneyDisease] = useState(false)
   const [liverDisease, setLiverDisease] = useState(false)
-  const [pregnant, setPregnant] = useState(false)
 
   const handleSubmit = event => {
     event.preventDefault()
@@ -70,19 +15,10 @@ function RiskFactors({handleRiskFactors}) {
       breathingDisease,
       cancer,
       diabetes,
-      heartDisease,
-      immunosuppressantDisease,
-      immunosuppressantDrug,
-      heartDiseaseAlgo,
-      immunosuppressantDiseaseAlgo,
-      immunosuppressantDrugAlgo,
       kidneyDisease,
-      liverDisease,
-      pregnant: pregnant ? '1' : '0'
+      liverDisease
     })
   }
-
-  const deriveAlgoValue = value => !(value === "0") // false si 0, true sinon
 
   return (
     <article className='step' id='risk-factors'>
@@ -91,35 +27,6 @@ function RiskFactors({handleRiskFactors}) {
         <form onSubmit={handleSubmit}>
           <div className='complement-infos'>
             <ul>
-              <li
-                style={{
-                  display: 'block'
-                }}
-              >
-                <RadioChoices
-                  icon='fa-heartbeat'
-                  name='heartDisease'
-                  value={heartDisease}
-                  onChange={value => {
-                    setHeartDisease(value)
-                    setHeartDiseaseAlgo(deriveAlgoValue(value))
-                  }}
-                  choices={[
-                    {
-                      title: 'Oui',
-                      value: '1'
-                    },
-                    {title: 'Non', value: '0'},
-                    {title: 'Je ne sais pas', value: '999'}
-                  ]}
-                >
-                  Avez-vous une hypertension artérielle mal équilibrée ?
-                  <br />
-                  Ou une maladie cardiaque ou vasculaire ?
-                  <br />
-                  Ou prenez-vous un traitement à visée cardiologique ?
-                </RadioChoices>
-              </li>
               <li>
                 <i className='fas fa-heartbeat' aria-hidden='true' />
                 <label>Êtes-vous diabétique ?</label>
@@ -169,67 +76,6 @@ function RiskFactors({handleRiskFactors}) {
                   checked={liverDisease}
                   onChange={() => setLiverDisease(!liverDisease)}
                 />
-              </li>
-              <li style={{display: 'block'}}>
-                <RadioChoices
-                  icon='fa-baby'
-                  name='pregnant'
-                  value={pregnant}
-                  onChange={value => setPregnant(value)}
-                  choices={[
-                    {
-                      title: 'Oui',
-                      value: '1'
-                    },
-                    {title: 'Non', value: '0'},
-                    {title: 'Non applicable', value: '888'}
-                  ]}
-                >
-                  Êtes-vous enceinte ?
-                </RadioChoices>
-              </li>
-              <li style={{display: 'block'}}>
-                <RadioChoices
-                  icon='fa-procedures'
-                  name='immunosuppressantDisease'
-                  value={immunosuppressantDisease}
-                  onChange={value => {
-                    setImmunosuppressantDisease(value)
-                    setImmunosuppressantDiseaseAlgo(deriveAlgoValue(value))}
-                  }}
-                  choices={[
-                    {
-                      title: 'Oui',
-                      value: '1'
-                    },
-                    {title: 'Non', value: '0'},
-                    {title: 'Je ne sais pas', value: '999'}
-                  ]}
-                >
-                  Avez-vous une maladie connue pour diminuer vos défenses
-                  immunitaires ?
-                </RadioChoices>
-              </li>
-              <li style={{display: 'block'}}>
-                <RadioChoices
-                  icon='fa-procedures'
-                  name='immunosuppressantDrug'
-                  value={immunosuppressantDrug}
-                  onChange={value => {
-                    setImmunosuppressantDrug(value)
-                    setImmunosuppressantDrugAlgo(deriveAlgoValue(value))
-                  }}
-                  choices={[
-                    {
-                      title: 'Oui',
-                      value: '1'
-                    },
-                    {title: 'Non', value: '0'},
-                    {title: 'Je ne sais pas', value: '999'}
-                  ]}
-                >
-                  Prenez-vous un traitement immunosuppresseur ?
-                </RadioChoices>
               </li>
             </ul>
           </div>
